Separate typings from Data.ts
