+++ src/lib/data/prisma.ts
@@ -1,0 +1,19 @@
+import { PrismaClient } from "@prisma/client";
+
+let prisma: PrismaClient;
+
+if (typeof window === "undefined") {
+  if (process.env.NODE_ENV === "production") {
+    prisma = new PrismaClient();
+  } else {
+    if (!global.prisma) {
+      global.prisma = new PrismaClient();
+    }
+
+    prisma = global.prisma;
+  }
+}
+
+
+// @ts-ignore
+export default prisma;
