+++ src/lib/global.d.ts
@@ -1,0 +1,5 @@
+import { PrismaClient } from '@prisma/client';
+
+declare global {
+  var prisma: PrismaClient;
+}
