+++ src/routes/posts/index.ts
@@ -1,4 +1,5 @@
-import { Data, type PostsResult } from "$lib/data/Data";
+import { Data } from "$lib/data/Data";
+import type { PostsResult } from "$lib/data/datatypes";
 import type NormalRequestHandler from "$lib/RequestHandler";
 
 let db = new Data();
