+++ src/routes/users/[id]/index.ts
@@ -1,4 +1,5 @@
-import { Data, type UserResult } from "$lib/data/Data";
+import { Data } from "$lib/data/Data";
+import type { UserResult } from "$lib/data/datatypes";
 import type NormalRequestHandler from "$lib/RequestHandler";
 
 let db = new Data();
